added base module in tf train and improved hierarchy of Simulator
