+++ batchglm/api/models/nb_glm.py
@@ -1,2 +1,2 @@
-from batchglm.models.nb_glm import InputData_NBGLM, Model, Simulator
+from batchglm.models.nb_glm import InputData, Model, Simulator
 from batchglm.train.tf.nb_glm import Estimator
