+++ batchglm/models/base_glm/utils.py
