+++ batchglm/models/nb_glm/external.py
@@ -1,8 +1,9 @@
-from batchglm.models import BasicEstimator, BasicInputData, BasicModel, BasicSimulator
-from batchglm.models.glm import parse_design, BasicGLM, InputData_GLM, INPUT_DATA_PARAMS
-import batchglm.utils.random as rand_utils
-import batchglm.data as data_utils
+from batchglm.models.base import _Estimator_Base
+from batchglm.models.base_glm import _InputData_GLM, _Model_GLM, _Simulator_GLM
+from batchglm.models.base_glm import INPUT_DATA_PARAMS
+from batchglm.models.base_glm import closedform_glm_mean, closedform_glm_var
 
+import batchglm.data as data_utils
+import batchglm.utils.random as rand_utils
+from batchglm.utils.numeric import weighted_mean, weighted_variance
 from batchglm.utils.linalg import groupwise_solve_lm
-from batchglm.utils.numeric import weighted_mean
-from batchglm.models.glm import closedform_glm_mean
