+++ batchglm/models/nb_glm/input.py
@@ -3,12 +3,12 @@ try:
 except ImportError:
     anndata = None
 
-from .external import InputData_GLM
+from .external import _InputData_GLM
 from .external import INPUT_DATA_PARAMS
 
 INPUT_DATA_PARAMS = INPUT_DATA_PARAMS.copy()
 
-class InputData_NBGLM(InputData_GLM):
+class InputData(_InputData_GLM):
     """
     Input data for Generalized Linear Models (GLMs) with negative binomial noise.
     """
