+++ batchglm/train/tf/base/estimator_graph.py
@@ -1,0 +1,15 @@
+import abc
+import tensorflow as tf
+
+
+class TFEstimatorGraph(metaclass=abc.ABCMeta):
+    graph: tf.Graph
+    loss: tf.Tensor
+    init_op: tf.Tensor
+    train_op: tf.Tensor
+    global_step: tf.Tensor
+
+    def __init__(self, graph=None):
+        if graph is None:
+            graph = tf.Graph()
+        self.graph = graph
