--- batchglm/train/tf/external.py
@@ -1,3 +1,0 @@
-import batchglm.pkg_constants as pkg_constants
-from batchglm.models.base import BasicEstimator
-import batchglm.utils.stats as stat_utils
