+++ batchglm/train/tf/nb_glm/external.py
@@ -1,11 +1,13 @@
 import batchglm.data as data_utils
 
-from batchglm.models.nb_glm import AbstractEstimator, XArrayEstimatorStore, InputData_NBGLM, Model
-import batchglm.models.nb_glm.utils as nb_glm_utils
+from batchglm.models.nb_glm import AbstractEstimator, XArrayEstimatorStore, InputData, Model
+from batchglm.models.nb_glm.utils import closedform_nb_glm_logmu, closedform_nb_glm_logphi
+
 
 import batchglm.train.tf.ops as op_utils
 import batchglm.train.tf.train as train_utils
 from batchglm.train.tf.base import TFEstimatorGraph, MonitoredTFEstimator
 
 import batchglm.utils.random as rand_utils
+from batchglm.utils.linalg import groupwise_solve_lm
 from batchglm import pkg_constants
