+++ batchglm/unit_test/test_nb_glm_extreme_values.py
@@ -6,13 +6,13 @@ import logging
 import numpy as np
 
 import batchglm.api as glm
-from batchglm.api.models.nb_glm import Simulator, Estimator, InputData_NBGLM
+from batchglm.api.models.nb_glm import Simulator, Estimator, InputData
 
-glm.setup_logging(verbosity="INFO", stream="STDOUT")
-logging.getLogger("tensorflow").setLevel(logging.INFO)
+glm.setup_logging(verbosity="ERROR", stream="STDOUT")
+logging.getLogger("tensorflow").setLevel(logging.ERROR)
 
 
-def estimate_byfeature(input_data: InputData_NBGLM, quick_scale):
+def estimate_byfeature(input_data: InputData, quick_scale):
     estimator = Estimator(
         input_data,
         batch_size=500,
@@ -32,7 +32,7 @@ def estimate_byfeature(input_data: InputData_NBGLM, quick_scale):
 
     return estimator
 
-def estimate_global(input_data: InputData_NBGLM, quick_scale):
+def estimate_global(input_data: InputData, quick_scale):
     estimator = Estimator(
         input_data,
         batch_size=500,
